+++ modulecheck-gradle/plugin/src/test/kotlin/modulecheck/gradle/TaskLifecycleTest.kt
@@ -1,0 +1,81 @@
+/*
+ * Copyright (C) 2021-2022 Rick Busarow
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package modulecheck.gradle
+
+import modulecheck.utils.child
+import modulecheck.utils.createSafely
+import org.gradle.testkit.runner.TaskOutcome.SUCCESS
+import org.junit.jupiter.api.Test
+
+class TaskLifecycleTest : BasePluginTest() {
+
+  @Test
+  fun `the modulecheck task should be invoked for the base plugin check task`() {
+
+    kotlinProject(":") {
+      buildFile {
+        """
+        plugins {
+          id("com.rickbusarow.module-check")
+          base
+        }
+        """
+      }
+
+      projectDir.child("settings.gradle.kts").createSafely()
+    }
+
+    shouldSucceed("check") {
+
+      task(":moduleCheck")!!.outcome shouldBe SUCCESS
+
+      tasks shouldBe listOf(
+        task(":moduleCheck"),
+        task(":check")
+      )
+    }
+  }
+
+  @Test
+  fun `the modulecheck task should be invoked for a late, manually created check task`() {
+
+    kotlinProject(":") {
+      buildFile {
+        """
+        plugins {
+          id("com.rickbusarow.module-check")
+        }
+
+        afterEvaluate {
+          task("check")
+        }
+        """
+      }
+
+      projectDir.child("settings.gradle.kts").createSafely()
+    }
+
+    shouldSucceed("check") {
+
+      task(":moduleCheck")!!.outcome shouldBe SUCCESS
+
+      tasks shouldBe listOf(
+        task(":moduleCheck"),
+        task(":check")
+      )
+    }
+  }
+}